# document/flowpilot-rest-api.openapi.yaml
openapi: 3.1.0
...
info:
  title: FlowPilot REST API
  version: 1.5.0
  license:
    name: Proprietary - Internal Use Only
    identifier: LicenseRef-Proprietary
  description: |
    FlowPilot 内网流程平台首版 REST API 契约。

    所有写操作由后端重新执行权限、状态与版本校验。可变资源使用整数
    `revision` 生成强 ETag；要求 `If-Match` 的接口在缺少请求头时返回 428，
    在版本过期时返回 412。创建及领域命令要求 `Idempotency-Key`，同一用户、
    路由作用域和键的相同请求会重放首次结果，不同请求体会返回 409。

    日期时间均为 ISO-8601 UTC。业务实体 ID 均由后端生成。审批、确认、字段
    修改、任务激活、流程状态、审计与邮件 Outbox 必须在同一业务事务中提交；
    SMTP 发送在事务提交后异步执行。Excel 文件只在浏览器中根据后端返回的
    已鉴权导出数据集生成，后端不生成、不保存 Excel 文件。

    浏览器只使用同源 HttpOnly 会话 Cookie，不保存访问令牌或业务数据。
servers:
  - url: /api/flowpilot/v1
    description: 与前端同源部署的内网 API
security:
  - cookieAuth: []
...
paths:
...
  /attachments:
    post:
      tags: [Attachments]
      operationId: uploadAttachment
      summary: 上传附件
      description: |
        仅创建暂存文件与元数据，不直接修改流程表单。正式引用必须随实例创建、表单保存、
        重新提交或任务决策原子提交。启用 PDF 展示的字段由服务端强制最多一个 attachmentId，
        新引用替换旧引用；旧物理文件仅在确认无其他引用后异步清理。富文本图片和视频使用
        purpose=rich-text-media 上传为后端托管媒体，正文仅保存附件标识和受认证内容地址。
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                definitionId:
                  type: string
                  description: 发起前暂存时必填，用于锁定字段策略
                versionId:
                  type: string
                  description: 发起前暂存时必填，不允许回退当前发布版本
                instanceId:
                  type: string
                  description: 已创建实例编辑时使用；与 definitionId/versionId 二选一
                fieldId:
                  type: string
                  description: 表单附件字段稳定标识
                purpose:
                  type: string
                  enum: [form-field, free-reply, rich-text-media]
      responses:
        "201":
          description: 附件已上传；普通附件处于暂存状态，富文本媒体处于可用状态
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "507":
          $ref: "#/components/responses/InsufficientStorage"
  /attachments/{attachmentId}:
    get:
      tags: [Attachments]
      operationId: getAttachment
      summary: 获取附件元数据
      parameters:
        - $ref: "#/components/parameters/AttachmentId"
      responses:
        "200":
          description: 附件元数据
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentDto"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Attachments]
      operationId: deleteStagedAttachment
      summary: 删除尚未被业务记录引用的暂存附件
      parameters:
        - $ref: "#/components/parameters/AttachmentId"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: 暂存附件已删除
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /attachments/{attachmentId}/content:
    get:
      tags: [Attachments]
      operationId: downloadAttachment
      summary: 按数据范围校验后下载或内联预览附件
      parameters:
        - $ref: "#/components/parameters/AttachmentId"
        - $ref: "#/components/parameters/Range"
        - name: disposition
          in: query
          schema:
            type: string
            enum: [inline, attachment]
            default: attachment
      responses:
        "200":
          description: 附件二进制内容
          headers:
            Accept-Ranges:
              description: 首版仅支持 bytes 单区间
              schema:
                type: string
                const: bytes
            Content-Length:
              schema:
                type: integer
                minimum: 0
            Content-Disposition:
              description: 安全编码后的原始文件名
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
            image/*:
              schema:
                type: string
                format: binary
            video/*:
              schema:
                type: string
                format: binary
        "206":
          description: 合法单区间附件内容
          headers:
            Accept-Ranges:
              schema:
                type: string
                const: bytes
            Content-Range:
              description: 实际返回的字节区间和完整长度
              schema:
                type: string
                example: bytes 0-1023/4096
            Content-Length:
              schema:
                type: integer
                minimum: 1
            Content-Disposition:
              description: 安全编码后的原始文件名
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "416":
          $ref: "#/components/responses/RangeNotSatisfiable"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: flowpilot_session
      description: 正式同源 HttpOnly 会话 Cookie，Path=/api/flowpilot；当前 HTTP 环境 Secure=false，启用 HTTPS 后必须为 true
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: GET 返回的强 ETag，例如 `"12"`
      schema:
        type: string
        pattern: '^"[0-9]+"$'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 客户端生成的 UUID；同作用域至少保留 24 小时
      schema:
        type: string
        minLength: 16
        maxLength: 100
...
    AttachmentId:
      name: attachmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Range:
      name: Range
      in: header
      required: false
      description: RFC 9110 bytes 单区间，例如 `bytes=0-1023`；首版拒绝多区间
      schema:
        type: string
        pattern: '^bytes=(?:[0-9]+-[0-9]*|-[0-9]+)$'
...
  headers:
    ETag:
      description: 由资源 revision 生成的强 ETag
      schema:
        type: string
        pattern: '^"[0-9]+"$'
...
  responses:
    BadRequest:
      description: 请求语法、查询范围或媒体格式参数错误
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Unauthorized:
      description: 未登录或会话失效
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Forbidden:
      description: 缺少页面/动作权限、流程资格或数据范围
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: 资源不存在，或为避免越权信息泄露而按不存在处理
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Conflict:
      description: 当前领域状态不允许该动作、幂等键冲突或字段并发冲突
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    PreconditionFailed:
      description: If-Match 已过期
      headers:
        ETag:
          $ref: "#/components/headers/ETag"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    UnprocessableEntity:
      description: 领域字段校验未通过
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    PreconditionRequired:
      description: 缺少必需的 If-Match
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    PayloadTooLarge:
      description: 文件超过字段或系统限制
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    UnsupportedMediaType:
      description: 文件内容类型不受支持
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    RangeNotSatisfiable:
      description: Range 无效、超出文件范围或包含首版不支持的多区间
      headers:
        Content-Range:
          description: 使用 `bytes */{完整长度}` 返回可用长度
          schema:
            type: string
            example: bytes */4096
        Accept-Ranges:
          schema:
            type: string
            const: bytes
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    InsufficientStorage:
      description: 附件卷无法在满足配置保留空间的前提下接受本次上传
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
  schemas:
...
    ProblemDetails:
      type: object
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri-reference
          example: /problems/task-already-completed
        title:
          type: string
          example: 任务已处理
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          $ref: "#/components/schemas/ProblemCode"
        traceId:
          type: string
        currentEtag:
          type: string
        retryAfterSeconds:
          type: integer
          minimum: 0
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssueDto"
        currentResource:
          description: 冲突处理所需的最小当前资源快照
          type: object
          additionalProperties: true
    ProblemCode:
      type: string
      enum:
        - BAD_REQUEST
        - AUTHENTICATION_REQUIRED
        - INVALID_CREDENTIALS
        - CSRF_VALIDATION_FAILED
        - PERMISSION_DENIED
        - DATA_SCOPE_DENIED
        - RESOURCE_NOT_FOUND
        - REVISION_MISMATCH
        - VALIDATION_FAILED
        - USER_ROLE_REQUIRED
        - IDEMPOTENCY_KEY_REUSED
        - IDEMPOTENCY_REQUEST_IN_PROGRESS
        - IMMUTABLE_BUILTIN_RESOURCE
        - RESOURCE_IN_USE
        - DELETE_BLOCKED
        - VERSION_NOT_EDITABLE
        - VERSION_PUBLISHED
        - VERSION_HAS_INSTANCES
        - VERSION_VALIDATION_FAILED
        - PUBLISH_POINTER_CHANGED
        - DEFINITION_DISABLED
        - DEFINITION_NOT_PUBLISHED
        - EXTERNAL_DEPENDENCY_INVALID
        - TASK_ALREADY_COMPLETED
        - ACTION_NOT_ALLOWED_FOR_NODE
        - REVIEW_COMMENT_REQUIRED
        - INSTANCE_CONTENT_LOCKED
        - INSTANCE_NOT_REJECTED
        - INSTANCE_ALREADY_CLOSED
        - REPEAT_EDIT_FORBIDDEN
        - NO_FIELD_CHANGES
        - FIELD_REVISION_CONFLICT
        - ATTACHMENT_NOT_STAGED
        - ATTACHMENT_IN_USE
        - ATTACHMENT_LIMIT_EXCEEDED
        - INVALID_ATTACHMENT_TYPE
        - PDF_ATTACHMENT_REQUIRED
        - ATTACHMENT_RANGE_NOT_SATISFIABLE
        - INSUFFICIENT_STORAGE
        - OUTBOX_MESSAGE_NOT_RETRYABLE
        - EXPORT_NOT_READY
        - EXPORT_EXPIRED
        - RATE_LIMITED
        - INTERNAL_SERVER_ERROR
        - DOMAIN_AUTHENTICATION_UNAVAILABLE
        - AUTHENTICATION_MODE_CONFLICT
        - IMPERSONATION_NOT_ALLOWED
        - IMPERSONATION_TARGET_INVALID
        - IMPERSONATION_ALREADY_ACTIVE
        - IMPERSONATION_SESSION_INVALID
        - IMPERSONATION_TARGET_NOT_FOUND
        - IF_MATCH_REQUIRED
        - PRECONDITION_REQUIRED
        - DEFINITION_NOT_FOUND
        - VERSION_NOT_FOUND
        - TASK_NOT_FOUND
        - INSTANCE_NOT_FOUND
        - INSTANCE_VIEW_FORBIDDEN
        - DEFINITION_NOT_LAUNCHABLE
        - LAUNCH_FORBIDDEN
        - USER_NOT_FOUND
        - ROLE_NOT_FOUND
        - DEPARTMENT_NOT_FOUND
        - POSITION_NOT_FOUND
        - WORKFLOW_GROUP_NOT_FOUND
        - ATTACHMENT_FILE_REQUIRED
        - ATTACHMENT_FILE_INVALID
        - ATTACHMENT_SCOPE_INVALID
        - ATTACHMENT_UPLOAD_FORBIDDEN
        - ATTACHMENT_INLINE_NOT_ALLOWED
        - INVALID_MULTIPART_BODY
        - MULTIPLE_RANGES_NOT_SUPPORTED
        - EMAIL_OUTBOX_NOT_FOUND
        - AUDIT_EVENT_NOT_FOUND
        - DATE_RANGE_INVALID
        - EXPORT_EMPTY_RESULT
        - EXPORT_NO_COLUMNS
        - EXPORT_ROW_LIMIT_EXCEEDED
    ValidationIssueDto:
      type: object
      required: [code, message]
      properties:
        path:
          type: string
          description: JSON Pointer 或稳定字段/节点路径
        code:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [error, warning]
          default: error
...
    UserRef:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        loginName:
          type: string
        email:
          type: string
          format: email
        departmentPath:
          type: string
...
    AttachmentStatus:
      type: string
      description: 普通业务接口可见状态；上传中、失败和已删除仅供内部恢复及运维使用
      enum: [staged, active, cleanup-pending]
    AttachmentDto:
      type: object
      required: [id, revision, originalName, sizeBytes, contentType, status, uploadedBy, uploadedAt]
      properties:
        id:
          type: string
          format: uuid
        revision:
          type: integer
          minimum: 1
        originalName:
          type: string
        sizeBytes:
          type: integer
          minimum: 0
        contentType:
          type: string
        sha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
        status:
          $ref: "#/components/schemas/AttachmentStatus"
        uploadedBy:
          $ref: "#/components/schemas/UserRef"
        uploadedAt:
          type: string
          format: date-time
        referencedBy:
          type: array
          readOnly: true
          items:
            type: object
            required: [aggregateType, aggregateId]
            properties:
              aggregateType:
                type: string
                enum: [process-instance, free-timeline-entry]
              aggregateId:
                type: string
                format: uuid
              fieldId:
                type: string
        contentUrl:
          type: string
          format: uri-reference
          readOnly: true
...